setAttributeNode and friends should not have optional argument https://bugs.webkit.org/show_bug.cgi?id=101631
Reviewed by Ojan Vafai.

Source/WebCore:

http://www.w3.org/TR/DOM-Level-3-Core/core.html#ID-887236154

Fix getAttributeNode, getAttributeNodeNS and removeAttributeNode to make the Attr
argument mandatory.

These used to throw DOMExceptions when an invalid type was passed instead of
TypeError which is also a spec violation.

Updated existing tests.

* bindings/scripts/CodeGeneratorV8.pm:
* dom/Element.cpp:
* dom/Element.idl:

LayoutTests:

Fix getAttributeNode, getAttributeNodeNS and removeAttributeNode to make the Attr
argument mandatory.

These used to throw DOMExceptions when an invalid type was passed instead of
TypeError which is also a spec violation.

* fast/dom/Element/attr-param-typechecking-expected.txt:
* fast/dom/Element/script-tests/attr-param-typechecking.js:


git-svn-id: svn://svn.chromium.org/blink/trunk@133944 bbb929c8-8fbe-4397-9dbb-9b2b20218538