[chromium] fake drag drop operations in the WebTestProxyBase instead of the WebViewHost https://bugs.webkit.org/show_bug.cgi?id=102495
Reviewed by Adam Barth.

When a drag operation is initiated in a layout test, we need to
simulate a corresponding drop event, otherwise the WebViewImpl will sit
there waiting for it. Instead of having the embedder taking care of
this, move this into the TestRunner library.

* DumpRenderTree/chromium/TestRunner/public/WebTestProxy.h:
(WebKit):
(WebTestProxyBase):
(WebTestRunner::WebTestProxy::startDragging):
* DumpRenderTree/chromium/TestRunner/src/WebTestProxy.cpp:
(WebTestRunner::WebTestProxyBase::startDragging):
(WebTestRunner):
* DumpRenderTree/chromium/WebViewHost.cpp:
* DumpRenderTree/chromium/WebViewHost.h:
(WebViewHost):


git-svn-id: svn://svn.chromium.org/blink/trunk@135006 bbb929c8-8fbe-4397-9dbb-9b2b20218538