Dump the tile cache extent in layout tests https://bugs.webkit.org/show_bug.cgi?id=102600
Reviewed by Anders Carlsson.

Source/WebCore:

When dumping tiled layer stats in layout tests, also dump the extent
of the tile grid. This will allow us to detect issues related
to zooming, which is not possible with the existing tile coverage rect,
which is dumped in layer (not tile) coordinates.

* platform/graphics/TiledBacking.h: Added tileGridExtent().
* platform/graphics/ca/GraphicsLayerCA.cpp:
(WebCore::GraphicsLayerCA::dumpAdditionalProperties): Dump the tile grid extent.
* platform/graphics/ca/mac/TileCache.h: tileCoverageRect() should be OVERRIDE.
* platform/graphics/ca/mac/TileCache.mm:
(WebCore::TileCache::getTileIndexRangeForRect):
(WebCore::TileCache::tileGridExtent): Return a rect with the size of
the grid as top,left width,height.

LayoutTests:

New results which contain tile cache extents.

* platform/mac/tiled-drawing/tile-coverage-after-scroll-expected.txt:
* platform/mac/tiled-drawing/tile-coverage-scroll-to-bottom-expected.txt:
* platform/mac/tiled-drawing/tile-coverage-slow-scrolling-expected.txt:
* platform/mac/tiled-drawing/tiled-drawing-zoom-expected.txt:
* platform/mac/tiled-drawing/tiled-drawing-zoom-scrolled-expected.txt:
* platform/mac/tiled-drawing/use-tiled-drawing-expected.txt:

git-svn-id: svn://svn.chromium.org/blink/trunk@135053 bbb929c8-8fbe-4397-9dbb-9b2b20218538