PDFPlugin: Support PDF form editing in <iframe>s https://bugs.webkit.org/show_bug.cgi?id=103287 <rdar://problem/10326917>
Reviewed by Dan Bernstein.

PDFPlugin form editing depends on having a Document that has the same
size and position as the plugin; this is true not only in the full-main-frame
case, but also in the <iframe> case. Therefore, we can enable form editing
for <iframe> PDFPlugins.

* WebProcess/Plugins/PDF/PDFPlugin.mm:
(WebKit::PDFPlugin::supportsForms): Support form editing whenever
our Document's widget is our PluginView (which will only be true in the
full-main-frame or <iframe> case).


git-svn-id: svn://svn.chromium.org/blink/trunk@136317 bbb929c8-8fbe-4397-9dbb-9b2b20218538