When animating mask-postion on a composited layer, element renders incorrectly https://bugs.webkit.org/show_bug.cgi?id=103700
Reviewed by Tim Horton.

Source/WebCore:

When a composited layer has a mask, we clipped the layer bounds to the
size of the mask. This caused problems when animating the mask origin,
which is considered a 'repaint' style change and so should not require
layout or compositing geometry changes.

Fix by ignoring the mask when computing the bounds of the compositing
layer.

Test: compositing/masks/mask-layer-size.html

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::localBoundingBox):
(WebCore::RenderLayer::calculateLayerBounds):
* rendering/RenderLayer.h:
(RenderLayer):
* rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::calculateCompositedBounds):

LayoutTests:

Test that dumps the layer tree for a composited layer with a mask.

* compositing/masks/mask-layer-size-expected.txt: Added.
* compositing/masks/mask-layer-size.html: Added.

git-svn-id: svn://svn.chromium.org/blink/trunk@136433 bbb929c8-8fbe-4397-9dbb-9b2b20218538