Extend StringHasher to take a stream of characters https://bugs.webkit.org/show_bug.cgi?id=104076 <rdar://problem/12811887>
Reviewed by Darin Adler.

We essentially move computeHash() to a member function of StringHasher. This allows clients of the
hasher to append strings to the general hash, without having to first concatenate the strings.

* wtf/StringHasher.h:
(WTF::StringHasher::addCharacters): Same as computeHash(), except also take into account the possibility
that there is a pending character to hash. If this is the case, we align the word boundary by consuming the
first character in data, adjust the length, move the data pointer forward one, and update the hasher's
pending character state. Also add an early return so that we are guaranteed at least one character in case
we run into this situation.
(WTF::StringHasher::computeHash): Refactor into addCharacters().

git-svn-id: svn://svn.chromium.org/blink/trunk@136695 bbb929c8-8fbe-4397-9dbb-9b2b20218538