[BlackBerry] Use midpoint for fixed position heuristic https://bugs.webkit.org/show_bug.cgi?id=105329
Patch by Andrew Lo <anlo@rim.com> on 2012-12-18
Reviewed by Rob Buis.
Internally reviewed by Joe Mason.

Use the midpoint of the layer when determining whether the
fixed-position layer is on the top or bottom half of the screen
rather than the top of the layer.
Internal PR 237073.

No new tests are needed.

* platform/graphics/blackberry/LayerRenderer.cpp:
(WebCore::LayerRenderer::updateLayersRecursive):

git-svn-id: svn://svn.chromium.org/blink/trunk@138056 bbb929c8-8fbe-4397-9dbb-9b2b20218538