[Mac] Add Build Phase to Check Headers for Inappropriate Macros (Platform.h macros) https://bugs.webkit.org/show_bug.cgi?id=104279
Reviewed by David Kilzer.

Source/JavaScriptCore:

Add a build phase to check the public JavaScriptCore headers for
inappropriate macros.

* JavaScriptCore.xcodeproj/project.pbxproj:

Source/WebCore:

Set the executable as an input path on the check-* build phases.

* WebCore.xcodeproj/project.pbxproj:

Source/WebKit:

Add a build phase to check the public WebKit headers for
inappropriate macros. Also set the executable as an input
path on the other check-* build phases.

* WebKit.xcodeproj/project.pbxproj:

Source/WebKit2:

Add a build phase to check the public WebKit2 headers for
inappropriate macros. Also set the executable as an input
path on the other check-* build phases.

* WebKit2.xcodeproj/project.pbxproj:

git-svn-id: svn://svn.chromium.org/blink/trunk@138065 bbb929c8-8fbe-4397-9dbb-9b2b20218538