Make CustomProtocolManager a WebProcessSupplement https://bugs.webkit.org/show_bug.cgi?id=105814
Reviewed by Dan Bernstein.

This converts CustomProtocolManager from a global singleton, to be a
supplement of the WebProcess. Since it also needs to act as a supplement
of the NetworkProcess, it adds an initialization function for the NetworkProcess
as well, while we await the arrival of the forthcoming NetworkProcessSupplement.
        
In addition, this moves the handling of registering and unregistering schemes
directly to the CustomProtocolManager, instead of requiring an unnecessary trip
in the ChildProcess.

* NetworkProcess/NetworkProcess.cpp:
(WebKit::NetworkProcess::NetworkProcess):
(WebKit::NetworkProcess::initializeNetworkProcess):
* NetworkProcess/NetworkProcess.h:
* NetworkProcess/NetworkProcess.messages.in:
* Shared/Network/CustomProtocols/CustomProtocolManager.h:
(CustomProtocolManager):
* Shared/Network/CustomProtocols/CustomProtocolManager.messages.in:
* Shared/Network/CustomProtocols/mac/CustomProtocolManagerMac.mm:
(+[WKCustomProtocol canInitWithRequest:]):
(-[WKCustomProtocol initWithRequest:cachedResponse:client:]):
(-[WKCustomProtocol startLoading]):
(-[WKCustomProtocol stopLoading]):
(WebKit::CustomProtocolManager::supplementName):
(WebKit::CustomProtocolManager::CustomProtocolManager):
(WebKit::CustomProtocolManager::initialize):
* UIProcess/WebContext.cpp:
(WebKit::WebContext::registerSchemeForCustomProtocol):
(WebKit::WebContext::unregisterSchemeForCustomProtocol):
* WebProcess/WebProcess.cpp:
(WebKit::WebProcess::WebProcess):
(WebKit::WebProcess::initializeWebProcess):
* WebProcess/WebProcess.h:
* WebProcess/WebProcess.messages.in:


git-svn-id: svn://svn.chromium.org/blink/trunk@138533 bbb929c8-8fbe-4397-9dbb-9b2b20218538