[BlackBerry] Allow defaultTabEventHandler to trigger on tab https://bugs.webkit.org/show_bug.cgi?id=106118
Reviewed by Yong Li.

In order to support tabbing through input elements, amongst
other things, we need to bipass the processing of the tab key as
'\t'. This will simply insert 4 spaces in the field. Remove tab
support from the key down map so that we don't translate it to
an editor command.

Internally reviewed by Mike Fenton.

* WebCoreSupport/EditorClientBlackBerry.cpp:
(WebCore):


git-svn-id: svn://svn.chromium.org/blink/trunk@138839 bbb929c8-8fbe-4397-9dbb-9b2b20218538