[Autotools] Add the AM_WEBKIT_FEATURE_CONDITIONAL macro https://bugs.webkit.org/show_bug.cgi?id=106576
Reviewed by Martin Robinson.

Replace the AC_CHECK_WEBKIT_FEATURE_ENABLED with the new macro.
The former was executing actions based on the passed-in feature being
enabled in the (possibly overridden) GNUmakefile.features.am, the actions
usually affecting the Automake conditional value that was set up afterwards.
The new macro does this directly, setting up an Automake conditional of the
same name as the feature that was checked.

* Source/autotools/webkitfeature.m4:


git-svn-id: svn://svn.chromium.org/blink/trunk@139327 bbb929c8-8fbe-4397-9dbb-9b2b20218538