GraphicsSurface: Canvas with WebGL content is painted off by one pixel https://bugs.webkit.org/show_bug.cgi?id=106446
When painting to TextureMapper the provided targetRect is in contents
coordinate system. This leads to an off by one pixel error when painting
outlines of 1px.
This needs to be taken into account in the transformation matrix
and in the TextureMapperGL::drawTexture call.

Test: fast/canvas/webgl/webgl-composite-modes.html

Reviewed by Noam Rosenthal.

* platform/graphics/surfaces/glx/GraphicsSurfaceGLX.cpp:
(WebCore::GraphicsSurface::platformPaintToTextureMapper):
* platform/graphics/surfaces/mac/GraphicsSurfaceMac.cpp:
(WebCore::GraphicsSurface::platformPaintToTextureMapper):
* platform/graphics/surfaces/win/GraphicsSurfaceWin.cpp:
(WebCore::GraphicsSurface::platformPaintToTextureMapper):

git-svn-id: svn://svn.chromium.org/blink/trunk@139738 bbb929c8-8fbe-4397-9dbb-9b2b20218538