[chromium] Make overlay layers slow-scrolling https://bugs.webkit.org/show_bug.cgi?id=108957
Patch by Sami Kyostila <skyostil@chromium.org> on 2013-02-05
Reviewed by James Robinson.

Since overlay layers get inserted on top of everything else, we must
mark them slow-scrolling to prevent all scroll input events to the rest
of the page from getting blocked. This is also more correct because
generally the overlay contents need to be repainted whenever the scroll
offset changes, and with this patch the painting happens in sync with
page scrolling.

git-svn-id: svn://svn.chromium.org/blink/trunk@141904 bbb929c8-8fbe-4397-9dbb-9b2b20218538